requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
jinja2==3.1.6
python-multipart==0.0.20
pydantic-settings==2.11.0
apscheduler==3.11.0
smartapi-python==1.5.5
pyotp==2.9.0
